apps/stockpilot/backend/requirements.txt
# 웹 프레임워크
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0

# 데이터베이스
sqlalchemy==2.0.23
psycopg2-binary==2.9.9
redis==5.0.1
alembic==1.12.1

# 데이터 처리 및 분석
pandas==2.2.3
numpy==1.25.2
scikit-learn==1.3.2
yfinance==0.2.22
ta==0.10.2
plotly==5.17.0

# API 및 데이터 수집
requests==2.31.0
aiohttp==3.9.0
beautifulsoup4==4.12.2
python-multipart==0.0.6
feedparser==6.0.10

# 보안 및 인증
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# 작업 큐 및 스케줄링
celery==5.3.4

# 로깅 및 모니터링
loguru==0.7.2
prometheus-client==0.19.0

# 개발 도구
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# 환경 변수 관리
python-dotenv==1.0.0

# 시간대 처리
pytz==2023.3
python-dateutil==2.8.2

# ML 및 텍스트 처리
joblib==1.3.2
textblob==0.17.1
openai==1.3.7

# 한국 시장 특화
holidays==0.36

# 암호화폐 데이터 (선택사항)
ccxt==4.1.52

# 이메일 알림
email-validator==2.1.0